Refractor

+++ gomcts/games.go
@@ -5,13 +5,13 @@ type GameResult float64
 
 // Action - interface representing entity that can be applied to a game state (generating the next game state)
 type Action interface {
-  ApplyTo(GameState) GameState
+	ApplyTo(GameState) GameState
 }
 
 // GameState - state of the game interface
 type GameState interface {
-  EvaluateGame() (GameResult, bool)
-  GetLegalActions() []Action
-  IsGameEnded() bool
-  NextToMove() int8
+	EvaluateGame() (GameResult, bool)
+	GetLegalActions() []Action
+	IsGameEnded() bool
+	NextToMove() int
 }
